bug : mauvais type dans la speed command
git-svn-id: svn://team-ard.com/ARP/branches/Protokrot_DVL_refactoringFaulhaber@434 c9b90f90-cbc1-11df-9d65-af01d29d43da

+++ arp_hml/src/hml/motor/Faulhaber3268Bx4.hpp
@@ -58,7 +58,7 @@ namespace arp_hml
     protected:
 
         ArdDs402::enum_DS402_state attrState;
-        int attrCommandedSpeed;
+        double attrCommandedSpeed;
         double attrPeriod;
 
         bool propInvertDriveDirection;
